+++ .clang-format
@@ -1,0 +1,4 @@
+BaseStyle: WebKit
+IndentWidth: 4
+TabWidth: 4
+UseTab: Never
